Refactor BatchID in its own module
