+++ core/src/models.rs
@@ -1,9 +1,11 @@
 pub mod account_state;
+pub mod batch_id;
 pub mod order;
 pub mod solution;
 pub mod tokens;
 
 pub use self::account_state::AccountState;
+pub use self::batch_id::BatchId;
 pub use self::order::Order;
 pub use self::solution::ExecutedOrder;
 pub use self::solution::Solution;
